+++ lib/dart_melty_soundfont.dart
@@ -1,6 +1,6 @@
 export 'dart:typed_data';
 
-export 'audio_renderer_ex.dart';
+export 'src/audio_renderer_ex.dart';
 export 'src/utils/array_int16.dart';
-export 'synthesizer.dart';
-export 'synthesizer_settings.dart';
+export 'src/synthesizer.dart';
+export 'src/synthesizer_settings.dart';
