+++ lib/dart_melty_soundfont_io.dart
@@ -1,0 +1,4 @@
+export 'dart_melty_soundfont.dart';
+
+export 'src/midi_file_io.dart';
+export 'src/soundfont_io.dart';
