+++ lib/src/modulation_envelope.dart
