+++ lib/src/modulator.dart
@@ -1,4 +1,4 @@
-import 'src/binary_reader.dart';
+import 'binary_reader.dart';
 
 class Modulator {
   // Since modulators will not be supported, we discard the data.
