+++ lib/src/region_ex.dart
@@ -1,9 +1,9 @@
 import 'oscillator.dart';
 import 'volume_envelope.dart';
 import 'modulation_envelope.dart';
-import 'src/lfo.dart';
-import 'src/utils/array_int16.dart';
-import 'src/instrument_region.dart';
+import 'lfo.dart';
+import 'utils/array_int16.dart';
+import 'instrument_region.dart';
 import 'region_pair.dart';
 import 'preset_region.dart';
 import 'soundfont_math.dart';
