+++ lib/src/soundfont_math.dart
