+++ lib/src/synthesizer_settings.dart
