+++ lib/src/volume_envelope.dart
